+++ src/mixer.cpp
@@ -11,7 +11,7 @@
 
 #include "../include/grape2d.h"
 
-void G2D_Engine::G2D_Mixer::setChannelVolume(int8_t volume, int channel){
+void G2D_Engine::G2D_Mixer::setChannelVolume(int volume, int channel){
     int v = (int)round(((float)volume/100.0f)*MIX_MAX_VOLUME);
 
     Mix_Volume(channel, v);
@@ -42,27 +42,27 @@ void G2D_Engine::G2D_Mixer::endChannel(int channel) {
     Mix_HaltChannel(channel);
 }
 
-void G2D_Engine::G2D_Mixer::endChannelTimed(int channel, uint32_t time_ms) {
+void G2D_Engine::G2D_Mixer::endChannelTimed(int channel, int time_ms) {
     Mix_ExpireChannel(channel, time_ms);
 }
 
-void G2D_Engine::G2D_Mixer::endChannelFadeOut(int channel, uint32_t time_ms) {
+void G2D_Engine::G2D_Mixer::endChannelFadeOut(int channel, int time_ms) {
     Mix_FadeOutChannel(channel, time_ms);
 }
 
-void G2D_Engine::G2D_Mixer::setPlanning(int channel, uint8_t left, uint8_t right) {
+void G2D_Engine::G2D_Mixer::setPlanning(int channel, int left, int right) {
     if(!Mix_SetPanning(channel, left, right)) {
         printf("Error on set planning. %s\n", Mix_GetError());
     }
 }
 
-void G2D_Engine::G2D_Mixer::setDistance(int channel, uint8_t distance) {
+void G2D_Engine::G2D_Mixer::setDistance(int channel, int distance) {
     if(!Mix_SetDistance(channel, distance)) {
         printf("Error on set distance. %s\n", Mix_GetError());
     }
 }
 
-void G2D_Engine::G2D_Mixer::setPosition(int channel, int16_t angle, uint8_t distance) {
+void G2D_Engine::G2D_Mixer::setPosition(int channel, int angle, int distance) {
     if(!Mix_SetPosition(channel, angle, distance)) {
         printf("Error on set position. %s\n", Mix_GetError());
     }
@@ -110,9 +110,9 @@ void G2D_Engine::G2D_Mixer::update() {
                 angle = angle * 180.0f / PI;
 
                 if (distance > _max_sound_distance) distance = _max_sound_distance;
-                auto d = (uint8_t) ((distance / (double) _max_sound_distance) * 255);
+                auto d = (int) ((distance / (double) _max_sound_distance) * 255);
 
-                G2D_Engine::instance->mixer->setPosition(i, (int16_t) angle, d);
+                G2D_Engine::instance->mixer->setPosition(i, (int) angle, d);
             }
             else{
                 _sound[i] = nullptr;
